Refactored StringUtils and HasLowerCase.py tests.
