+++ CodingPractice/PythonAssignments/cleancode/StringUtils.py
@@ -13,5 +13,5 @@ def num_of_strings_greater_than_n(n, list_of_str):
     return len([s for s in list_of_str if len(s) > n])
 
 
-def is_lower_case(string):
+def has_lower_case(string):
     return any(s.islower() for s in string)
